# Replace pthread and boost::thread with std::async