Fixes #466 and #464.