@lonvia 2015-11-08
@alex85k Could you run this on Windows to make sure it actually works as expected?