@pnorman 2015-11-08
node-persistent-cache-reader.cpp is now the only user of the HAVE_FORK define