@alex85k 2015-11-08
@lonvia  It's great, thatnk you!
Will run the tests with debug version now and then try some benchmarks.