@alex85k 2015-11-08
The usual tests seem to pass. (actually, there are same 3 failing muti-backend tests with master branch, but it is for another ticket).

However, the Python regression test (which is multithreaded, I guess), breaks each osm2pgsql with some "heap memory corruption" error deep inside geos call. Will investigate more and put the details here.