@alex85k 2015-11-08
I am afraid Python regression tests still do not work. There are access violation errors in async call:
![image](https://cloud.githubusercontent.com/assets/4124648/11021395/cc64fcc6-8661-11e5-9ae6-f0705ba7b1ec.png)