@alex85k 2015-11-08
Will post additional info if I find the exact variables and objects that are broken.