@alex85k 2015-11-08
It fails if I run with 4 threads only
```
osm2pgsql.exe RU-TY.osm.bz2 --number-processes 4
```
(1 or 2 processes => no errors).
And, as classical threading error, it differs for each run:
Run 1:
![image](https://cloud.githubusercontent.com/assets/4124648/11021872/523443f2-8671-11e5-8e71-fad6b3fc94cc.png)
Run 2:
![image](https://cloud.githubusercontent.com/assets/4124648/11021865/158e8188-8671-11e5-8616-5bb101fa92ec.png)
Run 3:
![image](https://cloud.githubusercontent.com/assets/4124648/11021883/90313520-8671-11e5-8806-2b9500691602.png)
The last was "vector iterator not dereferencable" exception in finding tag:
big screenshot is here https://db.tt/O5DcclDt .