@lonvia 2015-11-08
These errors seem to happen with stack-allocated data. So my uneducated guess would be: stack overflow. Is there a way to increase the stacksize for std::threads in MSVC, just to check if it runs through with more stack.