@alex85k 2015-11-09
Tried increasing stack size from 30Mb to 300Mb (/STACK linker option) - memory consumption is twice higher, but the errors are the same - always inside or after get_wkt_split . Maybe my geos 3.5.0 from git is buggy... 

Debbuging showed that the node and tag lists are small at the breaking point (5-8 node ways with ~4 tags) and coordinate sequences seem to be OK. I can not see any common data used. All the workers seem to be still running (tried to add  pritnf and breakpoint after w.get() in process_ways()), so there are no destruction-related errors too.