@lonvia 2015-11-09
Thanks, so we can exclude stack problems. It might be possible that your geos library is somehow not thread-safe (either because the C++ interface of geos doesn't use the thread-safe functions of geos or because it was built wrongly). I'm not sure how the tag list corruption would fit, although it is possible that geos corrupts other memory if it happens to use a thread-unsafe version of malloc.