@alex85k 2015-11-10
Intel tools finds nothing particular (few memory leaks in middle-ram, new ... as a parameters). 
The program does not crash or show critical error while running with analysis tools (it is slowing the threads down very much), so there are no new ideas... 

The only find: the errors does not appear if I do not call get_wkt_split or if I insert ``return nullptr`` before 
```c++
wkts->back().geom = writer.write(geom.get());
```
and after get_wkt_split call. But the errors resume (with some pif I move ``return nullptr`` after that line. It usually happens in geom.isValid() call  or inside writer.write.