@lonvia 2015-11-10
How about definining a global mutex in geometry_builder.cpp and adding a lock_guard around get_wkt_split (and potentially other functions that use writer).