@alex85k 2015-11-10
Thank you, mutex in all writers helped :) Running all tests & benchmarking Release to check...

Was it really using some common objects? Was the mutex really necessary or it is just a workaround for now?