@lonvia 2015-11-10
Jay, that's good news. Let me dig into the geos source code a bit and see if there isn't something to make the calls thread-safe. It's just strange that this problem never appeared on Linux, but I guess that's just the nature of multi-threading bugs.