@alex85k 2015-11-10
Tested: some speedup is possible even with mutex. 625s -> 425s (pending ways 98s -> 56s). Used 4 threads on moble Intel i7 Haswell 4200u (2 real cores).