@alex85k 2015-11-10
By the way, one lock_guard in  get_wkt_split  seemed to be enough...
https://github.com/alex85k/osm2pgsql/commits/replace-pthread-with-stdthread