@alex85k 2015-11-10
The tests passes with only one lock_guard. Strange result is that 1 processes is faster that 2 or 4 in that case (this time bulgaria-latest.osm.pbf was used). Time of processing pending ways increases from 34 to 43 s (1->2 processes).