@lonvia 2015-11-10
Those are odd numbers. With only one thread I would expect numbers to be approximately equal, no matter how many lock_guards there are.

I've added a fix for some variable shadowing on this branch. I doubt it makes a difference but could you give it a try anyways? I have the nagging feeling that somewhere in the code a freed pointer is accessed. Will stare at the code some more tomorrow.