@alex85k 2015-11-11
The lock_guard tests were not good because of different data formats used.

Tried with latest commit https://github.com/lonvia/osm2pgsql/commit/68e328d9962aa97212d01973e846e2a454088cb4 - the errors are the same for now :)