@alex85k 2015-11-11
By the way, one possible threading isuue:
![image](https://cloud.githubusercontent.com/assets/4124648/11085943/f35beb24-886d-11e5-9a70-c0397226de7f.png)
std::setlocale is set and restored from multiple threads.