@alex85k 2015-11-11
And when I deleted that locale switching code from geos (adding std::setlocale(LC_NUMERIC,"C") into main() ), the bug was gone. However, that is not a working solution...