@alex85k 2015-11-11
https://trac.osgeo.org/geos/ticket/260

It would be good to use stringstream.imbue for setting locale in geos instead of CLocalizer but I do not know how to make they accept it.