@lonvia 2015-11-11
Good catch. Sadly enough the geos code is so bad that there is no way to get around the setlocale() call. I wonder if we shouldn't simply just switch to WKB. It's CLocalizer free and also produces shorter strings.