@pnorman 2015-11-11
[a comment in output-pgsql.cpp](https://github.com/openstreetmap/osm2pgsql/blob/master/output-pgsql.cpp#L59-L70) discusses WKB, but for table_t::write_node a WKT is constructed by hand with boost::format so the comment might just be wrong. Something similar is done for ways.

I do recall issues with locale stuff not being thread-safe in Geos coming up previously, but can't find a reference. If this the issue, then it might be reproducible on linux by changing the locale to non-C before starting osm2pgsql.

There's a difference between WKB as text, EWKB as text, WKB, and EWKB, and I'm not sure which we'd want to send with either COPY or INSERT