@lonvia 2015-11-11
Postgres requires hex-encoded EWKB (with srid) iirc. Looks like geos is now able to produce that natively, see https://github.com/libgeos/libgeos/commit/cd9a22e7d6c8b7eab0813a9862c6a465bccce7f9