@lonvia 2015-11-11
Available since 3.0.0 (i.e. since 2007) if I read that commit right.